requirements.txt
camoufox[geoip]
playwright
pyyaml
rich
Pillow
schedule
